PMICDRV-99: Fix incorrect usages of PMIC_ST_ERR_INV_WDG_WINDOW
This return code is meant to be used when interacting with the Watchdog
module and it returns an error code indicating that a transaction
occurred outside of the expected window.

In some configuraton APIs, it was being used to indicate that the user
specified window configuation parameter was out of range; These should
be PMIC_ST_ERR_INV_PARAM instead.

Signed-off-by: Michael Leonard <m-leonard@ti.com>