PMICDRV-109: Remove "beginSequence" functions in WDG module
In code review we aligned that the functions dedicated to 'starting' a
watchdog sequence were not necessary. Fundamentally they were just
setting the watchdog mode (which can be done through Pmic_wdgSetMode())
and ensuring RETURN_LONGWIN and PWRHOLD bits were not set (which is the
default, and can be controlled through dedicated functions).

Signed-off-by: Michael Leonard <m-leonard@ti.com>