Add basic unit testing framework that can run on the Host OS
This adds a unit test framework that attempts to emulate the PMIC
register map to a 'good enough' extent to allow behavioral testing.
While this cannot fully replace testing on silicon, it does have the
following advantages:
- Tests are easier to implement
- Tests execute much faster as they are limited only by the procesing
  speed of the host machine
- Behavior that may not be easy to replicate and test on silicon can be
  more easily tested since we can arbitrarily control the register map.

There is still much work to do to make it actually 'emulate' anything,
but this is a good starting point.

The tests are executed using the [criterion][1] test
framework.

[1]: https://criterion.readthedocs.io

Signed-off-by: Michael Leonard <m-leonard@ti.com>