PMICDRV-98: Re-implement register lock control
Blackbird supports two types of register locking, CFG_REG_LOCK and
CNT_REG_LOCK. The former being a lock for configuration registers (when
it is set, they are read only), and the latter being a lock for the
timer control registers.

The previous implementation of these controls had separate config and
status structs for locking, and forced use of the structs for all
control/status reporting.

The revised implementation unifies the config/status into the same
struct and also provides helper functions for controlling each lock
individually.

Additionally, this commit addresses a potential bug in the previous
implemenation in which the driver may have proceeded with communication
and falsely reported success if the first byte sent had an error.

There are also various documentation improvments included in this
commit.

Signed-off-by: Michael Leonard <m-leonard@ti.com>