PMICDRV-100: Simplify project structure for single-device support
This commit more strictly enforces the separation between the purpose of
pmic.{c,h} and pmic_core.{c,h}. pmic_core is for control of core PMIC
features, while pmic is for core driver features (init/deinit).

This separation is enforced by creating a new pmic.c for implementing
the functions necessary for the functions included in pmic.h, instead of
implementing them in pmic_core.c.

Additionally, there are some functions that are needed throughout the
driver but not really needed by the customer, these were included in
pmic.h and marked as private. This commit moves them to pmic_types.{c,h}
instead, which is fully inteneded to be an internal only module.

Signed-off-by: Michael Leonard <m-leonard@ti.com>