Address Klockwork issues (MISRA-C, HIS, SA)
There are still some issues with higher cyclomatic complexity than
allowed by HIS metrics, but these are all limited to functions that are
fundamentally simple, and just appear to be complex with a bare
cyclomatic complexity calculation.

Additionally, there is an overflow risk in the WDG functions for
converting to/from ms & time codes for the watchdog window times, but we
have aligned on removing that functionality anyways.

Signed-off-by: Michael Leonard <m-leonard@ti.com>