pmic_wdg: Remove driver conversion of window duration
The driver was previously designed to require the user to configure the
watchdog windows using milliseconds, but this is error prone and
imprecise, and the user is essentially required to perform the
calculation anyways to verify they've chosen a value that makes sense.

We have decided to simplify the implementation and require that users
provide the raw code instead of a 'duration' in milliseconds.

Signed-off-by: Michael Leonard <m-leonard@ti.com>