PMICDRV-185: Add versions of Pmic_io{Tx,Rx}Byte which handle critical section
Add two new functions to the IO module which are duplicates of
Pmic_ioTxByte and Pmic_ioRxByte, but obtain and release critical
sections before/after performing the write/read, respectively.

These functions should be preferred for one-off reads and writes which
are not part of a larger sequence of sequential transactions.

Signed-off-by: Michael Leonard <m-leonard@ti.com>