PMICDRV-183: Potential PMIC_ST_INV_HANDLE error when Pmic_CoreHandle_t uses uninit memory
If the Pmic_CoreHandle_t type is created in an uninitialized memory
region the individual fields may not start out as 0. For most fields
this is okay as the fields are assigned using `=`, but the
`drvInitStatus` field is initialized using `|=` so will not end up with
the expected value unless it is explicitly initialized to 0.

This change ensures that this field starts out as `DRV_INIT_UNINIT`.

Signed-off-by: Michael Leonard <m-leonard@ti.com>