PMICDRV-210: Add support for pseudo-nINT
In systems where the nINT line is not connected to the processor, an
alternate method of detecting PMIC interrupt events must be used. As WD
Q&A servicing is a relatively frequent Processor/PMIC interaction, this
is a good time to do so.

To facilitate this, there is a bit in the WD_QA_CNT register which
indicates whether an interrupt is pending or not, and this can be
leveraged to detect interrupt events.

This commit adds a new callback function to the Pmic_CoreCfg_t
structure (and the handle) which allows uers to define a callback
function to be called when this interrupt bit is set.

Signed-off-by: Michael Leonard <m-leonard@ti.com>